localAiAssistant/requirements.txt
streamlit
streamlit-webrtc

langchain
langchain-community
langchain-core
langchain-openai

llava
ollama

pdfplumber
pypdf
pandas
numpy

scikit-learn
faiss-cpu

pyttsx3
openai-whisper
soundfile
pydub
pillow

tqdm
sentence-transformers
python-docx

watchdog
matplotlib

gtts
